requirements.txt
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.9.0
et-xmlfile==1.1.0
idna==3.6
keyboard==0.13.5
lxml==5.0.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
premailer==3.10.0
pycparser==2.21
PyPDF2==3.0.1
pypdfium2==4.25.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
yagmail==0.15.293
